Video render engine (#125)

+++ src/php/queries/give-up-from-competition.php
@@ -14,6 +14,7 @@
       ');
       $st->execute([$args->idComp, $args->idComp, $args->token]);
       $row = $st->fetch();
+      $user = $row['id'];
 
       if($row === false || $row['comps'] !== 1 || $row['pars'] !== 1){
         $this->err('data');
@@ -21,10 +22,33 @@
       }
 
       $st = $pdo->prepare('
+        select startDate
+        from Competition
+        where idComp = ?
+      ');
+      $st->execute([$args->idComp]);
+      $row = $st->fetch();
+
+      if($this->date >= $row['startDate']){
+        $this->err('compStartedGiveUp');
+        return;
+      }
+
+      $st = $pdo->prepare('
         delete from Participating
         where user = ? and comp = ?
       ');
-      $st->execute([$row['id'], $args->idComp]);
+      $st->execute([$user, $args->idComp]);
+
+      $st = $pdo->prepare('
+        update Competition
+        set currentUsers = currentUsers - 1
+        where idComp = ?
+      ');
+      $st->execute([
+        $args->idComp,
+      ]);
+
       $this->succ();
     }
   }
